refactor: Clean-up shared state for PostPage
