+++ src/components/PostPage.js
@@ -1,11 +1,24 @@
-import { useParams, Link } from 'react-router-dom';
+import { useParams, Link, useHistory } from 'react-router-dom';
 import { useContext } from 'react';
 import { DataContext } from '../context/DataContext';
+import api from '../api/posts';
 
 const PostPage = () => {
-  const { posts, handleDelete } = useContext(DataContext);
+  const { posts, setPosts } = useContext(DataContext);
   const { id } = useParams();
+  const history = useHistory();
   const post = posts.find((post) => post.id.toString() === id);
+
+  const handleDelete = async (id) => {
+    try {
+      api.delete(`/posts/${id}`);
+      const postsList = posts.filter((post) => post.id !== id);
+      setPosts(postsList);
+      history.push('/');
+    } catch (error) {
+      console.log(`Something went wrong: ${error.message}`);
+    }
+  };
 
   return (
     <main className='PostPage'>
